backend/pyproject.toml
[project]
name = "portada-backend"
version = "0.1.0"
description = "PortAda FastAPI Backend for data ingestion and analysis"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "PyYAML>=6.0.1",
    "py-portada-data-layer @ git+https://github.com/portada-git/py_portada_data_layer.git",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "pyjwt>=2.10.1",
]
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "hypothesis>=6.92.0",
    "pytest-mock>=3.12.0",
]
